fix: Allow starter kit env to work like SDK env
The starter kit is providing a single env var with JSON that
needs to be mapped to the standard SDK env vars.

Hopefully this is temporary code that can be removed some day,
but for now it allows the skit practices to work w/ SDK
practices with just a little bit of ugly code.